feat: functional build
Summary:
This is the initial release and seems to be operational for my needs.
PRs welcome if you want to expand the feature set.

Test Plan:
Ran the command on macos m1 pro and confirmed the video is converted
with decent color recreation from ps5 shares and an intro and outro.
Confirmed video starts uploading.
Confirmed videos are found in multiple levels of nesting
Confirmed that video files are uploading.
Confirmed the features function much the same way from inside a docker
container